Major revision to add dataset flexibility.
This revision uses array variables to allow multiple datasets to be used for dedupe and snapshots. This allows multiple datasets to be mounted into a common path for push to cloud services, but still dedupe and snapshot independently.

Some feature flags for rclone were also changed to increase reliability of transfer ETA for > 10,000 objects, and increase speed by using transfer 8 streams by default.