Add sorting option to jdupes.
The -O option preserves the sorting order by paths passed for evaluation. This helps maintain a consistent preserved file when the duplicate hardlinks are deleted, thereby not accidentally uploading multiple copies to the cloud providers unnecessarily. The option has been applied to both the hardlink operation and the delete-hardlink operation, though the former most likely does nothing (but also does not cause detriment).